requirements.txt
requests
flask
gunicorn
python-dotenv
numpy
transformers==4.40.1
torch==2.2.0
scikit-learn
